requirements.txt
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers

numpy>=1.16.2
scikit-learn>=0.20.3
matplotlib>=2.2.2
nibabel>=2.3.3
pillow>=5.4.1
torch==1.6
torchvision==0.7
psutil>=5.4.8
vtk>=8.2.0
pykeops==1.4.1
PyQt5
